Compiler fixes
