+++ src/App.tsx
@@ -3,7 +3,7 @@ import PipelineViewer from './components/PipelineViewer';
 import Tools from './components/Tools';
 import MapArea from './components/MapArea';
 import './App.css';
-import Store from './Store';
+import Store, { Outcomes } from './Store';
 import { MoveEastHandler, MoveNorthHandler, MoveSouthHandler, MoveWestHandler } from './actions/move-north';
 import { FeatureCollection } from 'geojson';
 import { ScaleUpHandler } from './actions/scale-track';
@@ -30,9 +30,9 @@ const App: React.FC = () => {
   const [outcomes, setOutcomes] = useState<{ [key: string]: any }>({});
   const [actions, setActions] = useState<BaseAction[]>([]);
 
-  const stateListener = (state: FeatureCollection | null) => {
+  const stateListener = (state: FeatureCollection | null, outcomes: Outcomes) => {
     setState(state);
-    setOutcomes(store.getOutcomes());
+    setOutcomes(outcomes);
   }
 
   const actionsListener = useCallback((actions: BaseAction[]) => {
